+++ src/main/java/com/keeper/homepage/domain/member/api/MemberController.java
@@ -2,13 +2,17 @@ package com.keeper.homepage.domain.member.api;
 
 import com.keeper.homepage.domain.member.application.MemberService;
 import com.keeper.homepage.domain.member.dto.request.ChangePasswordRequest;
+import com.keeper.homepage.domain.member.dto.response.MemberPointRankResponse;
 import com.keeper.homepage.domain.member.dto.response.MemberResponse;
 import com.keeper.homepage.domain.member.entity.Member;
 import com.keeper.homepage.global.config.security.annotation.LoginMember;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.PositiveOrZero;
 import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -44,5 +48,14 @@ public class MemberController {
       @RequestParam(required = false) String searchName
   ) {
     return ResponseEntity.ok(memberService.getMembersByRealName(searchName));
+  }
+
+  @GetMapping("/point-rank")
+  public ResponseEntity<Page<MemberPointRankResponse>> getPointRanks(
+      @RequestParam(defaultValue = "0") @PositiveOrZero int page,
+      @RequestParam(defaultValue = "10") @PositiveOrZero int size
+  ) {
+    Page<MemberPointRankResponse> response = memberService.getPointRanking(PageRequest.of(page, size));
+    return ResponseEntity.ok(response);
   }
 }
