+++ src/main/java/com/keeper/homepage/domain/member/application/MemberService.java
@@ -1,12 +1,17 @@
 package com.keeper.homepage.domain.member.application;
 
+import static com.keeper.homepage.domain.member.application.convenience.MemberFindService.VIRTUAL_MEMBER_ID;
+
 import com.keeper.homepage.domain.member.application.convenience.MemberFindService;
 import com.keeper.homepage.domain.member.dao.MemberRepository;
+import com.keeper.homepage.domain.member.dto.response.MemberPointRankResponse;
 import com.keeper.homepage.domain.member.dto.response.MemberResponse;
 import com.keeper.homepage.domain.member.entity.Member;
 import com.keeper.homepage.domain.member.entity.embedded.RealName;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,5 +37,10 @@ public class MemberService {
     return memberFindService.findAllByRealName(RealName.from(searchName))
         .map(MemberResponse::from)
         .toList();
+  }
+
+  public Page<MemberPointRankResponse> getPointRanking(Pageable pageable) {
+    return memberRepository.findAllByIdIsNotOrderByPointDesc(VIRTUAL_MEMBER_ID, pageable)
+        .map(MemberPointRankResponse::from);
   }
 }
