+++ src/test/java/com/keeper/homepage/domain/member/api/MemberApiTestHelper.java
@@ -1,7 +1,6 @@
 package com.keeper.homepage.domain.member.api;
 
 import static com.keeper.homepage.global.config.security.data.JwtType.ACCESS_TOKEN;
-import static com.keeper.homepage.global.restdocs.RestDocsHelper.field;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 
@@ -11,11 +10,11 @@ import org.springframework.restdocs.payload.FieldDescriptor;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.util.MultiValueMap;
 
-public class MemberRankApiTestHelper extends IntegrationTest {
+public class MemberApiTestHelper extends IntegrationTest {
 
   ResultActions callGetPointRankingApi(String accessToken, MultiValueMap<String, String> params)
       throws Exception {
-    return mockMvc.perform(get("/members/ranking/point")
+    return mockMvc.perform(get("/members/point-rank")
         .params(params)
         .cookie(new Cookie(ACCESS_TOKEN.getTokenName(), accessToken)));
   }
